Move stuff around to make more sens
