+++ commons.py
@@ -2,14 +2,20 @@ import yaml
 import logging
 
 import logger
-from definitions import CONFIG_PATH, CHATID_PATH
+
+from telegram.ext import ConversationHandler
+
+from definitions import CONFIG_PATH, CHATID_PATH, ADMIN_PATH, LANG_PATH
 
 config = yaml.safe_load(open(CONFIG_PATH, encoding="utf8"))
+
+lang = config["language"]
+transcript = yaml.safe_load(open(LANG_PATH, encoding="utf8"))
+transcript = transcript[lang]
 
 # Set up logging
 logLevel = logging.DEBUG if config.get("debugLogging", False) else logging.INFO
 logger = logger.getLogger("addarr.commons", logLevel, config.get("logToConsole", False))
-
 
 def generateServerAddr(app):
     try:
@@ -64,3 +70,35 @@ def checkId(update):
             return True
         else:
             return False
+
+def authentication(update, context):
+    chatid = update.effective_message.chat_id
+    with open(CHATID_PATH, "r") as file:
+        if(str(chatid) in file.read()):
+            context.bot.send_message(
+                chat_id=update.effective_message.chat_id,
+                text=transcript["Chatid already allowed"],
+            )
+            file.close()
+        else:
+            file.close()
+            password = update.message.text
+            if("/auth" in password):
+                password = password.replace("/auth ", "")
+            if password == config["telegram"]["password"]:
+                with open(CHATID_PATH, "a") as file:
+                    file.write(str(chatid) + "\n")
+                    context.bot.send_message(
+                        chat_id=update.effective_message.chat_id,
+                        text=transcript["Chatid added"],
+                    )
+                    file.close()
+                    return "added"
+            else:
+                logger.warning(
+                    f"Failed authentication attempt by [{update.message.from_user.username}]. Password entered: [{password}]"
+                )
+                context.bot.send_message(
+                    chat_id=update.effective_message.chat_id, text=transcript["Wrong password"]
+                )
+                return ConversationHandler.END # This only stops the auth conv, so it goes back to choosing screen
